Finished switching from iOS library project to a PCL library.
For the most part, this conversion was pretty straight forward. I did run into one complication in that I can't install RestSharp or the Newtonsoft via Nuget. To get around this, I used a PCL dll for Newtonsoft and abstracted the RestSharp functionality out of the PCL library.

Using MonkeyArms, I map a wrapper class for RestSharp to the IJSONWebClient the service code references.